Themes

+++ ip_cms/modules/standard/design/ThemeDownloader.php
@@ -48,14 +48,20 @@ tf1Tcb4xZFMMKDn/WwIDAQAB
 
         $this->extractZip(BASE_DIR . TMP_FILE_DIR . $themeTempFilename, BASE_DIR . THEME_DIR);
 
+        // TODOX patikrinti su Mangirdu ar verta pažeisti Law of Demeter
+        $fs = new  \Modules\administrator\system\Helper\FileSystem();
+        if (!$fs->makeWritable(BASE_DIR . THEME_DIR . $name, 0777)) {
+            throw new \Ip\CoreException('CHMOD failed.');
+        }
+
         unlink($archivePath);
     }
 
-    private function extractZip($archivePath, $destinationDir)
+    protected function extractZip($archivePath, $destinationDir)
     {
         if (class_exists('\\ZipArchive')) {
             $zip = new \ZipArchive();
-            if ($zip->open($archivePath) === TRUE) {
+            if ($zip->open($archivePath) === true) {
                 $zip->extractTo($destinationDir);
                 $zip->close();
             } else {
@@ -68,4 +74,5 @@ tf1Tcb4xZFMMKDn/WwIDAQAB
             }
         }
     }
+
 }
